// src/yeying/api/invitation/invitation.ts
export enum InvitationSceneEnum {
  /** INVITATION_SCENE_UNKNOWN - 未知编码，是一种占位符 */
  INVITATION_SCENE_UNKNOWN = 0,
  /** INVITATION_SCENE_USER - 用户邀请码，成为某个服务的用户，用于不完全开放的服务，只允许受邀使用该服务 */
  INVITATION_SCENE_USER = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface InvitationMetadata {
  /** 使用场景 */
  scene: InvitationSceneEnum;
  /** 邀请码 */
  code: string;
  /** 创建时间 */
  createdAt: string;
  /** 有效截止日期 */
  expiredAt: string;
  /** 邀请者身份ID */
  inviter: string;
  /** 受邀者身份ID，可以不填，表示邀请码可以给任何身份使用 */
  invitee: string;
  /** 邀请者签名 */
  signature: string;
}
// ...
export interface CreateInvitationResponse {
  header: MessageHeader | undefined;
  body: CreateInvitationResponseBody | undefined;
}

export interface CreateInvitationResponseBody {
  status: ResponseStatus | undefined;
  invitation: InvitationMetadata | undefined;
}
// ...
export const InvitationMetadata: MessageFns<InvitationMetadata> = {
  encode(message: InvitationMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.scene !== 0) {
      writer.uint32(8).int32(message.scene);
    }
    if (message.code !== "") {
      writer.uint32(18).string(message.code);
    }
    if (message.createdAt !== "") {
      writer.uint32(26).string(message.createdAt);
    }
    if (message.expiredAt !== "") {
      writer.uint32(34).string(message.expiredAt);
    }
    if (message.inviter !== "") {
      writer.uint32(42).string(message.inviter);
    }
    if (message.invitee !== "") {
      writer.uint32(50).string(message.invitee);
    }
    if (message.signature !== "") {
      writer.uint32(58).string(message.signature);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): InvitationMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseInvitationMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.scene = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.code = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.expiredAt = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.inviter = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.invitee = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.signature = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): InvitationMetadata {
    return {
      scene: isSet(object.scene) ? invitationSceneEnumFromJSON(object.scene) : 0,
      code: isSet(object.code) ? globalThis.String(object.code) : "",
      createdAt: isSet(object.createdAt) ? globalThis.String(object.createdAt) : "",
      expiredAt: isSet(object.expiredAt) ? globalThis.String(object.expiredAt) : "",
      inviter: isSet(object.inviter) ? globalThis.String(object.inviter) : "",
      invitee: isSet(object.invitee) ? globalThis.String(object.invitee) : "",
      signature: isSet(object.signature) ? globalThis.String(object.signature) : "",
    };
  },

  toJSON(message: InvitationMetadata): unknown {
    const obj: any = {};
    if (message.scene !== 0) {
      obj.scene = invitationSceneEnumToJSON(message.scene);
    }
    if (message.code !== "") {
      obj.code = message.code;
    }
    if (message.createdAt !== "") {
      obj.createdAt = message.createdAt;
    }
    if (message.expiredAt !== "") {
      obj.expiredAt = message.expiredAt;
    }
    if (message.inviter !== "") {
      obj.inviter = message.inviter;
    }
    if (message.invitee !== "") {
      obj.invitee = message.invitee;
    }
    if (message.signature !== "") {
      obj.signature = message.signature;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<InvitationMetadata>, I>>(base?: I): InvitationMetadata {
    return InvitationMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<InvitationMetadata>, I>>(object: I): InvitationMetadata {
    const message = createBaseInvitationMetadata();
    message.scene = object.scene ?? 0;
    message.code = object.code ?? "";
    message.createdAt = object.createdAt ?? "";
    message.expiredAt = object.expiredAt ?? "";
    message.inviter = object.inviter ?? "";
    message.invitee = object.invitee ?? "";
    message.signature = object.signature ?? "";
    return message;
  },
};
// ...
export const CreateInvitationResponse: MessageFns<CreateInvitationResponse> = {
  encode(message: CreateInvitationResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      CreateInvitationResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateInvitationResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateInvitationResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = CreateInvitationResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateInvitationResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? CreateInvitationResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: CreateInvitationResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = CreateInvitationResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateInvitationResponse>, I>>(base?: I): CreateInvitationResponse {
    return CreateInvitationResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateInvitationResponse>, I>>(object: I): CreateInvitationResponse {
    const message = createBaseCreateInvitationResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? CreateInvitationResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseCreateInvitationResponseBody(): CreateInvitationResponseBody {
  return { status: undefined, invitation: undefined };
}

export const CreateInvitationResponseBody: MessageFns<CreateInvitationResponseBody> = {
  encode(message: CreateInvitationResponseBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      ResponseStatus.encode(message.status, writer.uint32(10).fork()).join();
    }
    if (message.invitation !== undefined) {
      InvitationMetadata.encode(message.invitation, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateInvitationResponseBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateInvitationResponseBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = ResponseStatus.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.invitation = InvitationMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateInvitationResponseBody {
    return {
      status: isSet(object.status) ? ResponseStatus.fromJSON(object.status) : undefined,
      invitation: isSet(object.invitation) ? InvitationMetadata.fromJSON(object.invitation) : undefined,
    };
  },

  toJSON(message: CreateInvitationResponseBody): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = ResponseStatus.toJSON(message.status);
    }
    if (message.invitation !== undefined) {
      obj.invitation = InvitationMetadata.toJSON(message.invitation);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateInvitationResponseBody>, I>>(base?: I): CreateInvitationResponseBody {
    return CreateInvitationResponseBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateInvitationResponseBody>, I>>(object: I): CreateInvitationResponseBody {
    const message = createBaseCreateInvitationResponseBody();
    message.status = (object.status !== undefined && object.status !== null)
      ? ResponseStatus.fromPartial(object.status)
      : undefined;
    message.invitation = (object.invitation !== undefined && object.invitation !== null)
      ? InvitationMetadata.fromPartial(object.invitation)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}